+++ firestore-translate-text/functions/src/index.ts
@@ -38,19 +38,17 @@ const translate = new Translate({ projectId: process.env.PROJECT_ID });
 // Initialize the Firebase Admin SDK
 admin.initializeApp();
 
-logs.init();
+logs.init(config);
 
 export const fstranslate = functions.handler.firestore.document.onWrite(
   async (change): Promise<void> => {
-    logs.start();
-
+    logs.start(config);
     const { languages, inputFieldName, outputFieldName } = config;
 
     if (validators.fieldNamesMatch(inputFieldName, outputFieldName)) {
       logs.fieldNamesNotDifferent();
       return;
     }
-
     if (
       validators.fieldNameIsTranslationPath(
         inputFieldName,
@@ -120,35 +118,34 @@ const handleUpdateDocument = async (
   before: admin.firestore.DocumentSnapshot,
   after: admin.firestore.DocumentSnapshot
 ): Promise<void> => {
-  const inputAfter = extractInput(after);
   const inputBefore = extractInput(before);
+  const inputAfter = extractInput(after);
 
-  const inputHasChanged = inputAfter !== inputBefore;
-  if (
-    !inputHasChanged &&
-    inputAfter !== undefined &&
-    inputBefore !== undefined
-  ) {
-    logs.documentUpdatedUnchangedInput();
+  // If previous and updated documents have no input, skip.
+  if (inputBefore === undefined && inputAfter === undefined) {
+    logs.documentUpdatedNoInput();
     return;
   }
 
-  if (inputAfter) {
+  // If updated document has no string or object input, delete any existing translations.
+  if (typeof inputAfter !== "string" && typeof inputAfter !== "object") {
+    await updateTranslations(after, admin.firestore.FieldValue.delete());
+    logs.documentUpdatedDeletedInput();
+    return;
+  }
+
+  if (JSON.stringify(inputBefore) === JSON.stringify(inputAfter)) {
+    logs.documentUpdatedUnchangedInput();
+  } else {
     logs.documentUpdatedChangedInput();
     await translateDocument(after);
-  } else if (inputBefore) {
-    logs.documentUpdatedDeletedInput();
-    await updateTranslations(after, admin.firestore.FieldValue.delete());
-  } else {
-    logs.documentUpdatedNoInput();
   }
 };
 
-const translateDocument = async (
+const translateSingle = async (
+  input: string,
   snapshot: admin.firestore.DocumentSnapshot
 ): Promise<void> => {
-  const input: string = extractInput(snapshot);
-
   logs.translateInputStringToAllLanguages(input, config.languages);
 
   const tasks = config.languages.map(
@@ -173,11 +170,58 @@ const translateDocument = async (
       {}
     );
 
-    await updateTranslations(snapshot, translationsMap);
+    return updateTranslations(snapshot, translationsMap);
   } catch (err) {
     logs.translateInputToAllLanguagesError(input, err);
     throw err;
   }
+};
+
+const translateMultiple = async (
+  input: object,
+  snapshot: admin.firestore.DocumentSnapshot
+): Promise<void> => {
+  let translations = {};
+  let promises = [];
+
+  Object.entries(input).forEach(([input, value]) => {
+    config.languages.forEach((language) => {
+      promises.push(
+        () =>
+          new Promise(async (resolve) => {
+            logs.translateInputStringToAllLanguages(value, config.languages);
+
+            const output =
+              typeof value === "string"
+                ? await translateString(value, language)
+                : null;
+
+            if (!translations[input]) translations[input] = {};
+            translations[input][language] = output;
+
+            return resolve();
+          })
+      );
+    });
+  });
+
+  for (const fn of promises) {
+    if (fn) await fn();
+  }
+
+  return updateTranslations(snapshot, translations);
+};
+
+const translateDocument = async (
+  snapshot: admin.firestore.DocumentSnapshot
+): Promise<void> => {
+  const input: any = extractInput(snapshot);
+
+  if (typeof input === "object") {
+    return translateMultiple(input, snapshot);
+  }
+
+  await translateSingle(input, snapshot);
 };
 
 const translateString = async (