+++ jest.config.js
@@ -2,12 +2,11 @@ module.exports = {
   projects: [
     "<rootDir>/*/jest.config.js",
     "<rootDir>/*/functions/jest.config.js",
+    "<rootDir>/firestore-bigquery-export/scripts/*/jest.config.js",
   ],
   testPathIgnorePatterns: [
     ".*/bin/",
     ".*/lib/",
-    // ignore until existing tests migrated
-    ".*/firestore-bigquery-export/",
     ".*/firestore-counter/",
     // Ignoring otherwise tests duplicate due to Jest `projects`
     ".*/__tests__/.*.ts",