+++ storage-resize-images/functions/src/logs.ts
@@ -14,18 +14,23 @@
  * limitations under the License.
  */
 
+import { logger } from "firebase-functions";
 import config from "./config";
 
 export const complete = () => {
-  console.log("Completed execution of extension");
+  logger.log("Completed execution of extension");
 };
 
 export const noContentType = () => {
-  console.log(`File has no Content-Type, no processing is required`);
+  logger.log(`File has no Content-Type, no processing is required`);
+};
+
+export const gzipContentEncoding = () => {
+  logger.log("Images encoded with 'gzip' are not supported by this extension");
 };
 
 export const contentTypeInvalid = (contentType: string) => {
-  console.log(
+  logger.log(
     `File of type '${contentType}' is not an image, no processing is required`
   );
 };
@@ -34,7 +39,7 @@ export const unsupportedType = (
   unsupportedTypes: string[],
   contentType: string
 ) => {
-  console.log(
+  logger.log(
     `Image type '${contentType}' is not supported, here are the supported file types: ${unsupportedTypes.join(
       ", "
     )}`
@@ -42,81 +47,94 @@ export const unsupportedType = (
 };
 
 export const error = (err: Error) => {
-  console.error("Error when resizing image", err);
+  logger.error("Error when resizing image", err);
 };
 
 export const errorDeleting = (err: Error) => {
-  console.warn("Error when deleting temporary files", err);
+  logger.warn("Error when deleting temporary files", err);
 };
 
 export const failed = () => {
-  console.log("Failed execution of extension");
+  logger.log("Failed execution of extension");
 };
 
 export const imageAlreadyResized = () => {
-  console.log("File is already a resized image, no processing is required");
+  logger.log("File is already a resized image, no processing is required");
 };
 
 export const imageDownloaded = (remotePath: string, localPath: string) => {
-  console.log(`Downloaded image file: '${remotePath}' to '${localPath}'`);
+  logger.log(`Downloaded image file: '${remotePath}' to '${localPath}'`);
 };
 
 export const imageDownloading = (path: string) => {
-  console.log(`Downloading image file: '${path}'`);
+  logger.log(`Downloading image file: '${path}'`);
+};
+
+export const imageConverting = (
+  originalImageType: string,
+  imageType: string
+) => {
+  console.log(
+    `Converting image from type, ${originalImageType}, to type ${imageType}.`
+  );
+};
+
+export const imageConverted = (imageType: string) => {
+  console.log(`Converted image to ${imageType}`);
 };
 
 export const imageResized = (path: string) => {
-  console.log(`Resized image created at '${path}'`);
+  logger.log(`Resized image created at '${path}'`);
 };
 
 export const imageResizing = (path: string, size: string) => {
-  console.log(`Resizing image at path '${path}' to size: ${size}`);
+  logger.log(`Resizing image at path '${path}' to size: ${size}`);
 };
 
 export const imageUploaded = (path: string) => {
-  console.log(`Uploaded resized image to '${path}'`);
+  logger.log(`Uploaded resized image to '${path}'`);
 };
 
 export const imageUploading = (path: string) => {
-  console.log(`Uploading resized image to '${path}'`);
+  logger.log(`Uploading resized image to '${path}'`);
 };
 
 export const init = () => {
-  console.log("Initializing extension with configuration", config);
+  logger.log("Initializing extension with configuration", config);
 };
 
 export const start = () => {
-  console.log("Started execution of extension with configuration", config);
+  logger.log("Started execution of extension with configuration", config);
 };
 
 export const tempDirectoryCreated = (directory: string) => {
-  console.log(`Created temporary directory: '${directory}'`);
+  logger.log(`Created temporary directory: '${directory}'`);
 };
 
 export const tempDirectoryCreating = (directory: string) => {
-  console.log(`Creating temporary directory: '${directory}'`);
+  logger.log(`Creating temporary directory: '${directory}'`);
 };
 
 export const tempOriginalFileDeleted = (path: string) => {
-  console.log(`Deleted temporary original file: '${path}'`);
+  logger.log(`Deleted temporary original file: '${path}'`);
 };
 
 export const tempOriginalFileDeleting = (path: string) => {
-  console.log(`Deleting temporary original file: '${path}'`);
+  logger.log(`Deleting temporary original file: '${path}'`);
 };
 
 export const tempResizedFileDeleted = (path: string) => {
-  console.log(`Deleted temporary resized file: '${path}'`);
+  logger.log(`Deleted temporary resized file: '${path}'`);
 };
 
 export const tempResizedFileDeleting = (path: string) => {
-  console.log(`Deleting temporary resized file: '${path}'`);
+  logger.log(`Deleting temporary resized file: '${path}'`);
 };
 
 export const remoteFileDeleted = (path: string) => {
-  console.log(`Deleted original file from storage bucket: '${path}'`);
+  logger.log(`Deleted original file from storage bucket: '${path}'`);
 };
 
 export const remoteFileDeleting = (path: string) => {
-  console.log(`Deleting original file from storage bucket: '${path}'`);
+  logger.log(`Deleting original file from storage bucket: '${path}'`);
 };