+++ firestore-translate-text/CHANGELOG.md
@@ -1,3 +1,17 @@
+## Version 0.1.5
+
+docs - Improved documentation for input fields of type [`Map`](https://firebase.google.com/docs/firestore/manage-data/data-types#data_types).
+
+## Version 0.1.4
+
+feature - Support an input field of type [`Map`](https://firebase.google.com/docs/firestore/manage-data/data-types#data_types). If input is a `Map`, every value will be translated.
+
+feature - Add new Cloud Functions locations. For more information about locations and their pricing tiers, refer to the [location selection guide](https://firebase.google.com/docs/functions/locations).
+
+## Version 0.1.3
+
+feature - Update Cloud Functions runtime to Node.js 10.
+
 ## Version 0.1.2
 
 fixed - Fixed bug where target languages could not be reconfigured.
@@ -5,6 +19,7 @@ fixed - Fixed bug where target languages could not be reconfigured.
 ## Version 0.1.1
 
 fixed - Fixed bug where param validation failed when a single language was entered.
+
 fixed - Fixed "cold start" errors experienced when the extension runs after a period of inactivity (issue #48).
 
 ## Version 0.1.0